# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api-gateway service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/medals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method gets all medals",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MEDAL"
                ],
                "summary": "GET MEDALS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetMedalsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method creates a medal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MEDAL"
                ],
                "summary": "CREATE MEDAL",
                "parameters": [
                    {
                        "description": "Medal",
                        "name": "medal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMedalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Medal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/medals/filter": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method gets medals by filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MEDAL"
                ],
                "summary": "GET MEDALS BY FILTER",
                "parameters": [
                    {
                        "description": "Filter",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetMedalByFilterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetMedalByFilterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/medals/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method gets a medal by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MEDAL"
                ],
                "summary": "GET MEDAL BY ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetMedalByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method updates a medal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MEDAL"
                ],
                "summary": "UPDATE MEDAL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medal",
                        "name": "medal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMedalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Medal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This method deletes a medal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MEDAL"
                ],
                "summary": "DELETE MEDAL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteMedalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateMedalRequest": {
            "type": "object",
            "properties": {
                "athlete_id": {
                    "type": "string"
                },
                "country_id": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.MedalType"
                }
            }
        },
...
        "models.DeleteMedalResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.GetMedalByFilterRequest": {
            "type": "object",
            "properties": {
                "athlete_id": {
                    "type": "string"
                },
                "country_id": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.MedalType"
                }
            }
        },
        "models.GetMedalByFilterResponse": {
            "type": "object",
            "properties": {
                "medals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Medal"
                    }
                }
            }
        },
        "models.GetMedalByIdResponse": {
            "type": "object",
            "properties": {
                "athlete_id": {
                    "type": "string"
                },
                "country_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "integer"
                },
                "event_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.MedalType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.GetMedalsResponse": {
            "type": "object",
            "properties": {
                "medals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Medal"
                    }
                }
            }
        },
...
        "models.Medal": {
            "type": "object",
            "properties": {
                "athlete_id": {
                    "type": "string"
                },
                "country_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "integer"
                },
                "event_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.MedalType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.MedalType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "GOLD",
                "SILVER",
                "BRONZE"
            ]
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateMedalRequest": {
            "type": "object",
            "properties": {
                "athlete_id": {
                    "type": "string"
                },
                "country_id": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.MedalType"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}